requirements.txt
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
moexalgo==2.4.1
numpy==2.3.4
pandas==2.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
